+++ src/rp/Exercise/Ex3/Part2/Coord.java
@@ -1,31 +1,31 @@
 package rp.Exercise.Ex3.Part2;
 
 public class Coord {
-	private int x, y;
+	private byte x, y;
 
-	public Coord(int x, int y) {
+	public Coord(byte x, byte y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public byte getX() {
 		return this.x;
 	}
 
-	public int getY() {
+	public byte getY() {
 		return this.y;
 	}
 
-	public void setX(int x) {
+	public void setX(byte x) {
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(byte y) {
 		this.y = y;
 	}
 
 	public Coord getDelta(Coord c) {
-		return new Coord(c.x - this.x, c.y - this.y);
+		return new Coord((byte) (c.x - this.x), (byte) (c.y - this.y));
 	}
 
 	@Override
